Move omnilogic base entity to separate module
